annual-report-ai/requirements.txt
#just taking sure everyone uses the same packages;
#run this with pip install -r requirements.txt

fastapi
uvicorn
python-multipart
pdf2image
pytesseract
google-generativeai
python-dotenv
Pillow